LPS-104521 This should be an int

+++ modules/apps/item-selector/item-selector-taglib/src/test/java/com/liferay/item/selector/taglib/internal/display/context/GroupSelectorDisplayContextTest.java
@@ -90,7 +90,7 @@ public class GroupSelectorDisplayContextTest {
 		implements GroupItemSelectorProvider {
 
 		@Override
-		public long getGroupCount(
+		public int getGroupCount(
 			long companyId, long groupId, String keywords) {
 
 			return 3;
